refactor: made api more modular updated requirements.txt removed old app.py
