+++ backend/app/__init__.py
@@ -1,0 +1,30 @@
+from flask import Flask
+from flask_cors import CORS
+from flask_sqlalchemy import SQLAlchemy
+from app.config import Config
+
+# Initialize SQLAlchemy instance
+db = SQLAlchemy()
+
+def create_app():
+    """Create and configure the Flask application."""
+    app = Flask(__name__)
+    app.config.from_object(Config)
+
+    # Enable CORS with specific settings
+    CORS(app, resources={
+        r"/*": {
+            "origins": ["http://localhost:3000"],
+            "methods": ["GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"],
+            "allow_headers": ["Content-Type", "Authorization"]
+        }
+    })
+
+    # Initialize database with the app
+    db.init_app(app)
+
+    # Register blueprints
+    from app.routes.song_routes import song_bp
+    app.register_blueprint(song_bp)
+
+    return app
