+++ backend/app/models/song.py
@@ -1,17 +1,30 @@
-from sqlalchemy import Column, Integer, String, JSON, ForeignKey
-from sqlalchemy.orm import relationship
-from app.models.base import Base
+from app import db
+from datetime import datetime
 
-class Song(Base):
-    id = Column(Integer, primary_key=True, index=True)
-    title = Column(String, index=True)
-    key = Column(String)
-    tempo = Column(Integer)
-    time_signature = Column(String)
-    structure = Column(JSON)  # Store song structure (verses, choruses, etc.)
-    chords = Column(JSON)     # Store chord progressions
-    lyrics = Column(JSON)     # Store lyrics for different sections
-    user_id = Column(String, ForeignKey("user.id"))
-    
-    # Relationships
-    user = relationship("User", back_populates="songs") 
+class Song(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    title = db.Column(db.String(100), nullable=False)
+    key = db.Column(db.String(10), nullable=False)
+    tempo = db.Column(db.Integer, nullable=False)
+    time_signature = db.Column(db.String(10), nullable=False)
+    chords = db.Column(db.JSON, nullable=False)
+    structure = db.Column(db.JSON, nullable=False)
+    lyrics = db.Column(db.Text)
+    notes = db.Column(db.Text)
+    created_at = db.Column(db.DateTime, default=datetime.utcnow)
+    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
+
+    def to_dict(self):
+        return {
+            'id': self.id,
+            'title': self.title,
+            'key': self.key,
+            'tempo': self.tempo,
+            'timeSignature': self.time_signature,
+            'chords': self.chords,
+            'structure': self.structure,
+            'lyrics': self.lyrics,
+            'notes': self.notes,
+            'createdAt': self.created_at.isoformat(),
+            'updatedAt': self.updated_at.isoformat()
+        }
